+++ tests/PointInTime/HighResolutionTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Earth\PointInTime;
+namespace Tests\Innmind\TimeContinuum\PointInTime;
 
-use Innmind\TimeContinuum\Earth\PointInTime\HighResolution;
+use Innmind\TimeContinuum\PointInTime\HighResolution;
 use PHPUnit\Framework\TestCase;
 
 class HighResolutionTest extends TestCase